pyproject.toml
[project]
name = "gnn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "h5py>=3.15.1",
    "imageio>=2.37.2",
    "ipykernel>=7.1.0",
    "jupyterlab>=4.4.10",
    "matplotlib>=3.10.7",
    "optuna>=4.6.0",
    "rdkit>=2025.9.1",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "torch>=2.9.1",
    "torch-geometric>=2.7.0",
    "tqdm>=4.67.1",
    "umap-learn>=0.5.9.post2",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]

